requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
cffi==2.0.0
click==8.3.1
fastapi==0.123.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mutagen==1.47.0
numpy==2.3.5
pyalsaaudio==0.11.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyQt6==6.10.0
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.2
python-multipart==0.0.20
soundfile==0.13.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
